libwasi: fix build with old wasi-libc (and hopefully nuttx)
the nuttx bit has not been tested yet
